change 3:10

+++ app/src/main/java/com/odauday/ui/propertymanager/ActivityPropertyManager.java
@@ -1,5 +1,7 @@
 package com.odauday.ui.propertymanager;
 
+import android.app.AlertDialog;
+import android.app.AlertDialog.Builder;
 import android.app.ProgressDialog;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -9,7 +11,6 @@ import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.ImageView;
 import com.odauday.R;
 import com.odauday.data.local.cache.PrefKey;
 import com.odauday.data.local.cache.PreferencesHelper;
@@ -32,14 +33,17 @@ import timber.log.Timber;
  * Created by kunsubin on 4/18/2018.
  */
 
-public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyManagerBinding> implements PropertyManagerContract {
-    public static final String TAG=ActivityPropertyManager.class.getSimpleName();
+public class ActivityPropertyManager extends
+                                     BaseMVVMActivity<ActivityPropertyManagerBinding> implements
+                                                                                      PropertyManagerContract {
+    
+    public static final String TAG = ActivityPropertyManager.class.getSimpleName();
     
     @Inject
     PropertyManagerViewModel mPropertyManagerViewModel;
     @Inject
     PreferencesHelper mPreferencesHelper;
-    
+    AlertDialog.Builder mBuilderAlertDialog;
     private PropertyAdapter mPropertyAdapter;
     private ServiceUnavailableAdapter mServiceUnavailableAdapter;
     private EmptyPropertyAdapter mEmptyPropertyAdapter;
@@ -47,35 +51,48 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
     private Property mPropertyDelete;
     private List<Property> mProperties;
     private PopupMenu mPopupMenu;
-    
-    private PropertyAdapter.OnClickMenuListener mOnClickMenuListener=new OnClickMenuListener() {
+    private PropertyAdapter.OnClickMenuListener mOnClickMenuListener = new OnClickMenuListener() {
         @Override
         public void editProperty(Property property) {
-            Timber.tag(TAG).d("edit:"+property.getName());
+            Timber.tag(TAG).d("edit:" + property.getAddress());
         }
-    
+        
         @Override
         public void deleteProperty(Property property) {
-            Timber.tag(TAG).d("delete:"+property.getName());
-            if (property!=null){
-                mPropertyDelete=property;
-                mPropertyManagerViewModel.deleteProperty(property.getId());
+            Timber.tag(TAG).d("delete:" + property.getAddress());
+            
+            if (property != null) {
+                
+                mBuilderAlertDialog.setMessage(getString(R.string.message_delele_property));
+                mBuilderAlertDialog.setCancelable(true);
+                mBuilderAlertDialog.setIcon(getResources().getDrawable(R.drawable.ic_warning));
+                mBuilderAlertDialog
+                    .setPositiveButton(getString(R.string.txt_ok), (dialogInterface, i) -> {
+                        mPropertyDelete = property;
+                        mPropertyManagerViewModel.deleteProperty(property.getId());
+                    });
+                mBuilderAlertDialog
+                    .setNegativeButton(getString(R.string.txt_cancel), (dialogInterface, i) -> {
+                        dialogInterface.dismiss();
+                    });
+                AlertDialog alert11 = mBuilderAlertDialog.create();
+                alert11.show();
             }
         }
-    
+        
         @Override
         public void markTheEndProperty(Property property) {
-            Timber.tag(TAG).d("mark:"+property.getName());
+            Timber.tag(TAG).d("mark:" + property.getAddress());
         }
-    
+        
         @Override
         public void addExtendsTime(Property property) {
-            Timber.tag(TAG).d("add:"+property.getName());
+            Timber.tag(TAG).d("add:" + property.getAddress());
         }
-    
+        
         @Override
         public void reuseProperty(Property property) {
-            Timber.tag(TAG).d("reuse:"+property.getName());
+            Timber.tag(TAG).d("reuse:" + property.getAddress());
         }
     };
     
@@ -88,6 +105,7 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
     protected BaseViewModel getViewModel(String tag) {
         return mPropertyManagerViewModel;
     }
+    
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -95,22 +113,24 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
     }
     
     private void initView() {
-        mProgressDialog=new ProgressDialog(this);
+        mProgressDialog = new ProgressDialog(this);
         mProgressDialog.setCanceledOnTouchOutside(false);
         mProgressDialog.setMessage(getString(R.string.txt_progress));
-        
+        mBuilderAlertDialog = new Builder(this);
         mBinding.txtSearchBar.addTextChangedListener(new TextWatcher() {
             @Override
             public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
-        
+            
             }
+            
             @Override
             public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
                 Timber.tag(TAG).d(charSequence.toString());
-                if(mPropertyAdapter!=null){
+                if (mPropertyAdapter != null) {
                     mPropertyAdapter.filter(charSequence.toString());
                 }
             }
+            
             @Override
             public void afterTextChanged(Editable editable) {
             
@@ -127,7 +147,7 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
     private void getData() {
         mProgressDialog.show();
         mPropertyManagerViewModel.setPropertyManagerContract(this);
-        mPropertyManagerViewModel.getPropertyOfUser(mPreferencesHelper.get(PrefKey.USER_ID,""));
+        mPropertyManagerViewModel.getPropertyOfUser(mPreferencesHelper.get(PrefKey.USER_ID, ""));
     }
     
     @Override
@@ -151,20 +171,21 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
                 }
             }
         });
-        mPropertyAdapter=new PropertyAdapter();
+        mPropertyAdapter = new PropertyAdapter();
         mPropertyAdapter.setOnClickMenuListener(mOnClickMenuListener);
-        mBinding.recycleViewProperties.setLayoutManager(new GridLayoutManager(this,1));
+        mBinding.recycleViewProperties.setLayoutManager(new GridLayoutManager(this, 1));
         mBinding.recycleViewProperties.setNestedScrollingEnabled(false);
-        mServiceUnavailableAdapter=new ServiceUnavailableAdapter();
-        mEmptyPropertyAdapter=new EmptyPropertyAdapter();
+        mServiceUnavailableAdapter = new ServiceUnavailableAdapter();
+        mEmptyPropertyAdapter = new EmptyPropertyAdapter();
         
     }
+    
     @Override
     public void loading(boolean isLoading) {
         Timber.tag(TAG).d("loading");
-        if(isLoading){
+        if (isLoading) {
             mProgressDialog.show();
-        }else {
+        } else {
             mProgressDialog.dismiss();
         }
     }
@@ -172,25 +193,25 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
     @Override
     public void onSuccess(Object object) {
         Timber.tag(TAG).d("success");
-        List<Property> list=(List<Property>) object;
-        if(list!=null&&list.size()>0){
-            mProperties=list;
-            if(mBinding.recycleViewProperties.getAdapter() instanceof PropertyAdapter){
+        List<Property> list = (List<Property>) object;
+        if (list != null && list.size() > 0) {
+            mProperties = list;
+            if (mBinding.recycleViewProperties.getAdapter() instanceof PropertyAdapter) {
             
-            }else {
+            } else {
                 mBinding.recycleViewProperties.setAdapter(mPropertyAdapter);
             }
             mPropertyAdapter.setData(list);
             
-        }else {
+        } else {
             mBinding.recycleViewProperties.setAdapter(mEmptyPropertyAdapter);
         }
     }
     
     @Override
     public void onFailure(Exception ex) {
-        Timber.tag(TAG).d("error: "+ex.getMessage());
-    
+        Timber.tag(TAG).d("error: " + ex.getMessage());
+        
         String message;
         if (ex instanceof RetrofitException) {
             message = getString(R.string.message_service_unavailable);
@@ -201,7 +222,8 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
         }
         SnackBarUtils.showSnackBar(mBinding.propertyManager, message);
     }
-    public void onClickBack(View view){
+    
+    public void onClickBack(View view) {
         this.finish();
     }
     
@@ -209,25 +231,25 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
     public void onSuccessDeleteProperty(Object object) {
         Timber.tag(TAG).d("On Success delete");
         mBinding.txtSearchBar.setText("");
-        if(mProperties!=null&&mProperties.size()>0){
-            if(mPropertyDelete!=null){
+        if (mProperties != null && mProperties.size() > 0) {
+            if (mPropertyDelete != null) {
                 mProperties.remove(mPropertyDelete);
-                if(mBinding.recycleViewProperties.getAdapter() instanceof PropertyAdapter){
-        
-                }else {
+                if (mBinding.recycleViewProperties.getAdapter() instanceof PropertyAdapter) {
+                
+                } else {
                     mBinding.recycleViewProperties.setAdapter(mPropertyAdapter);
                 }
                 mPropertyAdapter.setData(mProperties);
             }
         }
-        MessageResponse messageResponse=(MessageResponse) object;
+        MessageResponse messageResponse = (MessageResponse) object;
         SnackBarUtils.showSnackBar(mBinding.propertyManager, messageResponse.getMessage());
     }
     
     @Override
     public void onErrorDeleteProperty(Object object) {
         Timber.tag(TAG).d("On Error delete");
-        Exception ex=(Exception) object;
+        Exception ex = (Exception) object;
         String message;
         if (ex instanceof RetrofitException) {
             message = getString(R.string.message_service_unavailable);
@@ -237,8 +259,9 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
         }
         SnackBarUtils.showSnackBar(mBinding.propertyManager, message);
     }
-    public void onClickMore(View view){
-        mPopupMenu = new PopupMenu(this,view);
+    
+    public void onClickMore(View view) {
+        mPopupMenu = new PopupMenu(this, view);
         mPopupMenu.getMenuInflater()
             .inflate(R.menu.menu_sort_property_manager, mPopupMenu.getMenu());
         mPopupMenu.setOnMenuItemClickListener(item -> {
@@ -247,8 +270,9 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
         });
         mPopupMenu.show();
     }
+    
     private void handlerClickItem(MenuItem item) {
-        switch (item.getItemId()){
+        switch (item.getItemId()) {
             case R.id.ascending:
                 Timber.tag(TAG).d("Ascending Click");
                 ascendingProperty();
@@ -257,47 +281,47 @@ public class ActivityPropertyManager extends BaseMVVMActivity<ActivityPropertyMa
                 Timber.tag(TAG).d("Descending Click");
                 descendingProperty();
                 break;
-                default:
-                    break;
+            default:
+                break;
         }
     }
     
     private void ascendingProperty() {
-        List<Property> list=mProperties;
-        Comparator<Property> comparator=(property1, property2) ->{
-            return property1.getName().compareTo(property2.getName());
+        List<Property> list = mProperties;
+        Comparator<Property> comparator = (property1, property2) -> {
+            return property1.getAddress().compareTo(property2.getAddress());
         };
-        if(list!=null&&list.size()>0){
+        if (list != null && list.size() > 0) {
             
-            Collections.sort(list,comparator);
+            Collections.sort(list, comparator);
             
-            if(mBinding.recycleViewProperties.getAdapter() instanceof PropertyAdapter){
-        
-            }else {
+            if (mBinding.recycleViewProperties.getAdapter() instanceof PropertyAdapter) {
+            
+            } else {
                 mBinding.recycleViewProperties.setAdapter(mPropertyAdapter);
             }
             mPropertyAdapter.setData(list);
-        }else {
+        } else {
             mBinding.recycleViewProperties.setAdapter(mEmptyPropertyAdapter);
         }
     }
     
     private void descendingProperty() {
-        List<Property> list=mProperties;
-        Comparator<Property> comparator=(property1, property2) ->{
-            return property2.getName().compareTo(property1.getName());
+        List<Property> list = mProperties;
+        Comparator<Property> comparator = (property1, property2) -> {
+            return property2.getAddress().compareTo(property1.getAddress());
         };
-        if(list!=null&&list.size()>0){
-        
-            Collections.sort(list,comparator);
-        
-            if(mBinding.recycleViewProperties.getAdapter() instanceof PropertyAdapter){
+        if (list != null && list.size() > 0) {
             
-            }else {
+            Collections.sort(list, comparator);
+            
+            if (mBinding.recycleViewProperties.getAdapter() instanceof PropertyAdapter) {
+            
+            } else {
                 mBinding.recycleViewProperties.setAdapter(mPropertyAdapter);
             }
             mPropertyAdapter.setData(list);
-        }else {
+        } else {
             mBinding.recycleViewProperties.setAdapter(mEmptyPropertyAdapter);
         }
     }
